feat: 리뷰 0.5점 별점과 작성 가능 여부 추가

+++ src/main/java/guesthouse/review/dto/response/ReviewResponse.java
@@ -16,7 +16,7 @@ public record ReviewResponse(
         Long userId,
         String authorName,
         String authorProfileImageUrl,
-        int rating,
+        double rating,
         String content,
         List<String> imageUrls,
         boolean isMine,
@@ -36,7 +36,7 @@ public record ReviewResponse(
                 .userId(review.getUserId())
                 .authorName(authorName)
                 .authorProfileImageUrl(authorProfileImageUrl)
-                .rating(review.getRating())
+                .rating(review.getRating().doubleValue())
                 .content(review.getContent())
                 .imageUrls(imageUrls)
                 .isMine(review.getUserId().equals(requestUserId))
